Missing sys reg handling (#228)
A fix for handling missing system registers in the aarch64 systemRegisterMap_ map. A missing entry will return a -1 and a decoded instruction accessing an unmapped system register will raise a new UnmappedSysReg fatal exception.

+++ src/lib/arch/aarch64/Architecture.cc
@@ -290,7 +290,9 @@ void Architecture::updateSystemTimerRegisters(RegisterFileSet* regFile,
 }
 
 simeng::Register Architecture::getPCCreg() const {
-  return {RegisterType::SYSTEM, getSystemRegisterTag(ARM64_SYSREG_PMCCNTR_EL0)};
+  return {
+      RegisterType::SYSTEM,
+      static_cast<uint16_t>(getSystemRegisterTag(ARM64_SYSREG_PMCCNTR_EL0))};
 }
 
 }  // namespace aarch64
